refactor: context를 별도 디렉토리로 분리
