+++ src/components/Information.jsx
@@ -1,7 +1,7 @@
-import React from "react";
-import styled from "styled-components";
-import { useTodoContext } from "./TodoContextProvider";
-import WarningModal from "../atomicComponents/WarningModal";
+import React from 'react';
+import styled from 'styled-components';
+import { useTodoContext } from '../context/TodoContextProvider';
+import WarningModal from '../atomicComponents/WarningModal';
 
 const Typography = styled.div`
   font-size: 2rem;
@@ -13,7 +13,9 @@ const Information = () => {
   return (
     <>
       {loading && <Typography>로딩중..</Typography>}
-      {warningVisible && <WarningModal>네트워크 에러가 발생했습니다</WarningModal>}
+      {warningVisible && (
+        <WarningModal>네트워크 에러가 발생했습니다</WarningModal>
+      )}
     </>
   );
 };
