+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
-import React from "react";
-import Welcome from "../components/Welcome";
-import Information from "../components/Information";
-import { useTodoContext } from "../components/TodoContextProvider";
+import React from 'react';
+import Welcome from '../components/Welcome';
+import Information from '../components/Information';
+import { useTodoContext } from '../context/TodoContextProvider';
 
 const Home = () => {
   const { loading, warningVisible } = useTodoContext();
